+++ database-observability/examples/exadata-database-service/providers.tf
@@ -1,0 +1,17 @@
+# Copyright (c) 2026 Contributors to oci-dbman-opsi.
+# Licensed under the Universal Permissive License v 1.0.
+
+terraform {
+  required_version = ">= 1.5.0, < 2.0.0"
+
+  required_providers {
+    oci = {
+      source  = "oracle/oci"
+      version = ">= 6.0.0, < 9.0.0"
+    }
+  }
+}
+
+provider "oci" {
+  region = var.region
+}